+++ CHANGELOG.md
@@ -5,6 +5,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+### Changed
+- `random` mode added to `decor` function
+- `both` parameter added to `decor` function
 ## [4.9] - 2020-10-16
 ### Added
 - 10 new fonts